Title: 산타의 선물 공장, Time: 153ms, Memory: 9MB, Status: Passed - CodeTree

+++ 240920/산타의 선물 공장/santa-gift-factory.cpp
@@ -81,8 +81,8 @@ void q_100() {
         }
     }
 }
-int q_200() {
-    int ret = 0;
+long long q_200() {
+    long long ret = 0;
     int w;
     cin >> w;
     for (int i = 0; i < m; ++i) {
